requirements.txt
flask==2.3.3
python-docx==0.8.11
PyPDF2==3.0.1
nltk==3.8.1
reportlab==4.0.7
python-dotenv==1.0.0
werkzeug==2.3.7
